@@ -21,11 +21,7 @@ setup(
     url="https://github.com/opspresso/toast-cli",
     packages=find_packages(),
     include_package_data=True,
-    package_data={
-        "toast_cli": ["VERSION"],
-        "toast": ["../VERSION"],
-    },
-    python_requires=">=3.6",
+    python_requires=">=3.9",
     install_requires=[
         "click",
         "rich",
@@ -37,10 +33,10 @@ setup(
     },
     classifiers=[
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
-        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+        "Programming Language :: Python :: 3.12",
         "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
         "Operating System :: OS Independent",
         "Environment :: Console",
